services/account_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyodbc
python-dotenv
bcrypt